+++ chaosreliably/controls/experiment.py
@@ -25,7 +25,6 @@ from logzero import logger
 
 from chaosreliably import RELIABLY_HOST, get_session
 from chaosreliably.activities.pauses import reset as reset_activity_pause
-from chaosreliably.types import AutoPause
 
 __all__ = ["configure_control"]
 
@@ -394,74 +393,17 @@ def configure_control(
     event_registry: EventHandlerRegistry,
     exp_id: str,
     org_id: str,
-    autopause: Optional[AutoPause] = None,
     configuration: Configuration = None,
     secrets: Secrets = None,
     **kwargs: Any,
 ) -> None:
     logger.debug("Configure Reliably's experiment control")
-
-    if autopause:
-        amend_experiment_for_autopauses(experiment, autopause)
-
     event_registry.register(ReliablyHandler(org_id, exp_id, experiment))
 
 
 ###############################################################################
 # Private functions
 ###############################################################################
-def amend_experiment_for_autopauses(
-    experiment: Experiment, autopause: AutoPause
-) -> None:
-    method = experiment.get("method")
-    if method and "method" in autopause:
-        p = autopause["method"]
-        after_actions = p.get("after_actions", True)
-        after_probes = p.get("after_probes", False)
-        pause_duration = p.get("pause_duration", 0)
-
-        activities = method[:]
-        for index, activity in enumerate(activities):
-            index = method.index(activity)
-            if after_probes and activity["type"] == "probe":
-                method.insert(index + 1, make_pause(pause_duration))
-            elif after_actions and activity["type"] == "action":
-                method.insert(index + 1, make_pause(pause_duration))
-
-    ssh_probes = experiment.get("steady-state-hypothesis", {}).get("probes")
-    if ssh_probes and "steady-state-hypothesis" in autopause:
-        p = autopause["steady-state-hypothesis"]
-        pause_duration = p.get("pause_duration", 0)
-
-        activities = ssh_probes[:]
-        for index, activity in enumerate(activities):
-            index = method.index(activity)
-            ssh_probes.insert(index + 1, make_pause(pause_duration))
-
-    rollbacks = experiment.get("rollbacks")
-    if rollbacks and "rollbacks" in autopause:
-        p = autopause["rollbacks"]
-        pause_duration = p.get("pause_duration", 0)
-
-        activities = rollbacks[:]
-        for index, activity in enumerate(activities):
-            index = method.index(activity)
-            rollbacks.insert(index + 1, make_pause(pause_duration))
-
-
-def make_pause(pause_duration: int = 0) -> Activity:
-    return {
-        "type": "action",
-        "name": f"reliably-pause-{secrets.token_hex(4)}",
-        "provider": {
-            "type": "python",
-            "module": "chaosreliably.activities.pauses",
-            "func": "pause_execution",
-            "arguments": {"duration": pause_duration},
-        },
-    }
-
-
 def make_user_pause(
     pause_duration: int = 0, username: str = "", user_id: str = ""
 ) -> Activity: